+++ Plugins/UnityPurchasing/iOS/SK2/Extensions/StringExtensions.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, visionOS 1.0, *)
 extension String {
 
     func toCString() -> UnsafePointer<CChar> {