+++ Plugins/UnityPurchasing/iOS/SK2/Models/ProductDefinition.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, visionOS 1.0, *)
 public struct ProductDefinition: Decodable {
     let id: String
     let storeSpecificId: String