+++ Plugins/UnityPurchasing/iOS/SK2/Responses/ResponseProtocol.swift
@@ -1,5 +1,6 @@
 /**
  This protocol provides interface for any responses from the StoreKits API call
  */
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, visionOS 1.0, *)
 public protocol ResponseProtocol {
 }